// backup_hanghand_20260113_223534/src/types/domain.ts
export interface RefCode {
  codeId: string; // e.g. '1010000'
  parentId?: string;
  type: 'INDUSTRY' | 'CATEGORY' | 'STATUS' | 'UNIT' | 'COMMUNITY_NODE';
  zhName: string;
  enName?: string;
  extraData?: Record<string, any>;
  sortOrder?: number;
}